python-backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
orjson
python-dotenv
pytest
httpx